strava webhok validation

+++ database/migrations/2022_09_28_144255_alter_users_table_add_strava_field.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('user', function (Blueprint $table) {
-            $table->integer('strava_athlete_id')->nullable();
+            $table->unsignedBigInteger('strava_athlete_id')->nullable();
             $table->string('strava_token', 255)->nullable();
             $table->string('strava_refresh_token', 255)->nullable();
             $table->integer('strava_expires_at')->nullable();
